Create SourceDirBuilder for separation of concerns
