+++ src/translator.py
@@ -3,6 +3,7 @@ import polib
 
 from Helpers import create_pot_file, create_pot_file_from_keyed, create_languagedata_xml_file, create_logger, merge_compendium
 from Parser import parse_arguments
+from source_dir_builder import SourceDirBuilder
 
 class Translator:
     def __init__(self):
@@ -17,13 +18,11 @@ class Translator:
         if os.path.exists('RimTranslate.log'):
             os.remove('RimTranslate.log')
 
-
     def start(self):
         if self.args.source_dir:
-            self.build_source_dir()
+            SourceDirBuilder(self.args, self.compendium, self.logger).build_source_dir()
         if self.args.output_dir:
             self.build_output_dir()
-
 
     def build_compendium(self):
         if self.compendium is not None:
@@ -46,70 +45,6 @@ class Translator:
                     compendium += create_pot_file_from_keyed(full_filename, self.args.source_dir, self.args.compendium, True)
 
         return compendium
-
-
-    def build_source_dir_defs(self):
-        self.build_source_dir_generic(
-            'DefInjected',
-            os.path.join(self.args.source_dir, 'Defs', '')
-        )
-
-
-    def build_source_dir_keyed(self):
-        self.build_source_dir_generic(
-            'Keyed',
-            os.path.join(self.args.source_dir, 'Languages/English/Keyed', '')
-        )
-
-    def build_source_dir(self):
-        self.logger.info('Beginning to generate PO-files')
-        self.build_source_dir_defs()
-        self.build_source_dir_keyed()
-
-
-
-    def build_source_dir_generic(self, category_name, source_dir, compendium_mode=False):
-        self.logger.info('Generating PO-files from %s' % category_name)
-
-        if not os.path.isdir(source_dir):
-            self.logger.error('%s is not a directory' % source_dir)
-            quit()
-
-        for root, dirs, files in os.walk(source_dir):
-            for file in files:
-                if file.endswith('.xml'):
-                    full_filename = os.path.join(root, file)
-                    self.logger.info("Processing " + full_filename)
-                    file_dir = full_filename.split(source_dir, 1)[1]
-
-                    if category_name == 'DefInjected':
-                        # Replace Defs to Def (https://github.com/winterheart/RimTranslate/issues/1)
-                        file_dir = file_dir.replace("Defs", "Def")
-
-                    pot = create_pot_file(category_name, full_filename, self.args.source_dir, self.args.compendium, compendium_mode)
-                    pofilename = os.path.join(self.args.po_dir, category_name, file_dir)
-                    pofilename += '.po'
-
-                    if os.path.exists(pofilename):
-                        self.logger.info("Updating PO file " + pofilename)
-                        po = polib.pofile(pofilename)
-                        po.merge(pot)
-                    else:
-                        # Is there some useful info?
-                        if len(pot) > 0:
-                            directory = os.path.dirname(pofilename)
-                            if not (os.path.exists(directory)):
-                                self.logger.info("Creating directory " + directory)
-                                os.makedirs(directory)
-                            self.logger.info("Creating PO file " + pofilename)
-                        po = pot
-
-                    if self.args.compendium:
-                        po = merge_compendium(self.compendium, po)
-
-                    if len(po):
-                        po.save(pofilename)
-
 
     def build_output_dir(self):
         self.logger.info('Beginning to generate DefInjected files')
